Reestructuring files and folders
