+++ src/Telemetry/Library/DataContracts/TelemetryPropertyType.cs
@@ -1,0 +1,18 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.Quantum.Telemetry
+{
+    /// <summary>
+    /// Allowed property types for the telemetry events
+    /// </summary>
+    public enum TelemetryPropertyType
+    {
+        String,
+        DateTime,
+        Long,
+        Double,
+        Guid,
+        Boolean,
+    }
+}
