+++ src/Telemetry/Library/OutOfProcess/OutOfProcessCommandType.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+namespace Microsoft.Quantum.Telemetry.OutOfProcess
+{
+    internal enum OutOfProcessCommandType
+    {
+        SetContext,
+        LogEvent,
+        Quit,
+    }
+}
